fix: remove abstract class and use a Map<String,Object> as key
