+++ src/main/java/io/kestra/plugin/aws/dynamodb/GetItem.java
@@ -11,7 +11,6 @@ import lombok.experimental.SuperBuilder;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 import software.amazon.awssdk.services.dynamodb.model.GetItemRequest;
 
-import java.util.HashMap;
 import java.util.Map;
 
 @SuperBuilder
@@ -28,8 +27,8 @@ import java.util.Map;
             title = "Get an item from its key.",
             code = {
                 "tableName: \"persons\"",
-                "keyName: \"id\"",
-                "keyValue; \"1\""
+                "key: ",
+                "   id: \"1\""
             }
         )
     }
@@ -37,22 +36,16 @@ import java.util.Map;
 public class GetItem extends AbstractDynamoDb implements RunnableTask<GetItem.Output> {
 
     @Schema(
-        title = "The DynamoDB item key name."
+        title = "The DynamoDB item key.",
+        description = "The DynamoDB item key. It's a map of string -> object."
     )
     @PluginProperty
-    private String keyName;
-
-    @Schema(
-        title = "The DynamoDB key value."
-    )
-    @PluginProperty(dynamic = true)
-    private String keyValue;
+    private Map<String, Object>  key;
 
     @Override
     public Output run(RunContext runContext) throws Exception {
         try (var dynamoDb = client(runContext)) {
-            Map<String, AttributeValue> key = new HashMap<>();
-            key.put(keyName, AttributeValue.fromS(runContext.render(keyValue)));
+            Map<String, AttributeValue> key = valueMapFrom(getKey());
 
             var getRequest = GetItemRequest.builder()
                 .tableName(runContext.render(this.getTableName()))
@@ -60,7 +53,7 @@ public class GetItem extends AbstractDynamoDb implements RunnableTask<GetItem.Ou
                 .build();
 
             var response = dynamoDb.getItem(getRequest);
-            var row = DynamoDbUtils.objectMapFrom(response.item());
+            var row = objectMapFrom(response.item());
             return Output.builder().row(row).build();
         }
     }
