+++ src/test/java/io/kestra/plugin/aws/dynamodb/GetItemTest.java
@@ -21,8 +21,7 @@ class GetItemTest extends AbstractDynamoDbTest {
             .accessKeyId(localstack.getAccessKey())
             .secretKeyId(localstack.getSecretKey())
             .tableName("persons")
-            .keyName("id")
-            .keyValue("1")
+            .key(Map.of("id", "1"))
             .build();
 
         createTable(runContext, get);
